+++ lib/providers/version_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-const String appVersion = '1.3.0';
+const String appVersion = '1.1.0';
 
 final appVersionProvider = Provider<String>((ref) => appVersion);